pyproject.toml
[tool.poetry]
name = "blitz"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.scripts]
blitz = "blitz.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
pyyaml = "^6.0.1"
pydantic = "^2.5.3"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
fastapi = "^0.108.0"
fastapi-crudrouter = "^0.8.6"
fastui = "0.4.0"
ipython = "^8.19.0"
openai = "^1.6.1"
nicegui = "^1.4.8"
jinja2 = "^3.1.3"
pick = "^2.2.0"
starlette-admin = {extras = ["sqla"], version = "^0.13.1"}
types-pyyaml = "^6.0.12.12"
alembic = {git = "https://github.com/mde-pach/alembic.git"}
semver = "^3.0.2"
loguru = "^0.7.2"
httpx = "^0.26.0"

[tool.poetry.group.doc.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.3"
mkdocs-markdownextradata-plugin = "^0.2.5"
mkdocs-redirects = "^1.2.1"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
mkdocs-typer = "^0.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"